level-8 refactor with OOP
Slave methods broken down into 3 classes:

*Ui
*Storage
*Parser

functionality unchanged
