+++ .gitignore
@@ -15,6 +15,5 @@ bin/
 
 /text-ui-test/ACTUAL.TXT
 text-ui-test/EXPECTED-UNIX.TXT
-username.txt
 *.class
 savefile.txt
