--- src/main/java/slaveFiles/NoUserDataException.java
@@ -1,8 +1,0 @@
-package slaveFiles;
-
-public class NoUserDataException extends Exception {
-    public NoUserDataException(String errorMsg) {
-        super(errorMsg);
-    }
-
-}
