+++ src/main/java/slaveFiles/Storage.java
@@ -1,0 +1,155 @@
+package slaveFiles;
+
+import java.io.BufferedWriter;
+import java.io.IOException;
+import java.io.FileWriter;
+import java.io.File;
+import java.io.FileNotFoundException;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeParseException;
+
+import java.util.List;
+import java.util.Scanner;
+
+public class Storage {
+    private String filePath;
+    private List<Task> list;
+
+    public Storage(List<Task> list, String filePath) {
+        this.list = list;
+        this.filePath = filePath;
+    }
+
+    /**
+     * converts the List<Task> to a string format and writes it to the savefile "./src/main/data/savefile.txt"
+     * every line contains only 1 task
+     * string format is as per the return value of toString() method of the respective task
+     * creates a new file at "./src/main/data" called "savefile.txt" in the event of a missing save file
+     */
+    public void save() {
+        try {
+            StringBuilder sb = new StringBuilder();
+            for (Task t : list) {
+                sb.append(t.save());
+                sb.append("\n");
+            }
+            BufferedWriter writer = new BufferedWriter(new FileWriter(filePath));
+            writer.write(sb.toString());
+            writer.close();
+//            System.out.println("saved");
+        } catch (IOException ioe) {
+            System.out.println("Save failed");
+            System.out.println(ioe.getMessage());
+        }
+    }
+
+    /**
+     * attempts to load the pre saved tasks from the save file "./src/main/data/savefile.txt"
+     * if file does not exist at the path, does nothing
+     * in the event of corrupted save file, skips the line
+     * save file format: as per toString() function of each task, with one task per line
+     */
+    public void load() {
+        try {
+            Scanner sc = new Scanner(new File(filePath));
+            int success = 0;
+            int failed = 0;
+            while (sc.hasNextLine()) {
+                try {
+                    String task = sc.nextLine();
+                    char taskType = task.charAt(1);
+                    char taskCompleted = task.charAt(4);
+                    int firstSpacePos = task.indexOf(" ");
+                    boolean completed = false;
+                    if (taskCompleted == ']') {
+                        // do nothing
+                    } else if (taskCompleted == 'X') {
+                        completed = true;
+                    } else {
+                        throw new InvalidSaveFileFormatException("invalid completed status");
+                    }
+                    // identify the type of task:
+                    switch (taskType) {
+                        case 'T':
+                            list.add(new Todo(completed, task.substring(firstSpacePos + 1)));
+                            success++;
+                            break;
+                        case 'D':
+                            // not sure how to get rid of error here for string formatting
+                            String[] deadlineArray = task.split(" \\(by: ");
+                            String deadlineName = deadlineArray[0].substring(firstSpacePos + 1);
+                            String by = deadlineArray[1].substring(0, deadlineArray[1].length() - 1);
+                            list.add(new Deadline(completed, deadlineName, LocalDate.parse(by)));
+                            success++;
+                            break;
+                        case 'E':
+                            // not sure how to get rid of error here for string formatting
+                            String[] eventArray = task.split(" \\(from: ");
+                            String eventName = eventArray[0].substring(firstSpacePos + 1);
+                            String[] eventDetails = eventArray[1].split(" to: ");
+                            String from = eventDetails[0];
+                            String to = eventDetails[1].substring(0, eventDetails[1].length() - 1);
+                            list.add(new Event(completed, eventName, LocalDate.parse(from), LocalDate.parse(to)));
+                            success++;
+                            break;
+                        default:
+                            throw new InvalidSaveFileFormatException("invalid Task type");
+                    }
+                } catch (InvalidSaveFileFormatException | DateTimeParseException
+                         | IndexOutOfBoundsException | InvalidChronologicalOrderException e) {
+                    // DateTimeParseException for incorrect LocalDate format
+                    // InvalidSaveFileFormatException, IndexOutOfBoundsException for incorrect Task format
+                    failed++;
+                    System.out.println("error in event save file format: " + e.getMessage());
+                }
+            }
+            if (failed > 0) {
+                System.out.println("Deleting buggy save file");
+                try {
+                    BufferedWriter writer = new BufferedWriter(new FileWriter("./src/main/data/savefile.txt"));
+                    writer.write("");
+                    writer.close();
+                    System.out.println("Buggy save file deleted");
+                    if (success > 0) {
+                        System.out.println("Saving readable contents");
+                        save();
+                    }
+                } catch (IOException e) {
+                    System.out.println("Failed to purge buggy save file");
+                }
+            }
+            if (success == 0) {
+                System.out.println("No valid save file found");
+            }
+        } catch (FileNotFoundException fnfe) {
+            System.out.println("no save file found");
+        }
+    }
+
+    /**
+     * marks the task as completed by changing the boolean value to true
+     *
+     * @param s is the task index in String format
+     * @throws IllegalArgumentException in the event that event index is out of the range of the task list
+     */
+    private void markAsDone(String s) throws IllegalArgumentException {
+        try {
+            int i = Integer.parseInt(s);
+
+            if (i < 1 || i > list.size()) {
+                System.out.println("You don't have a task number " + i);
+                return;
+            }
+            Task t = list.get(i - 1);
+            t.completed();
+            save();
+            System.out.println("Finally doing something useful with your life eh...");
+            System.out.println(t);
+        } catch (NumberFormatException nfe) {
+            System.out.println("That's not a task number");
+        }
+    }
+
+
+}
