+++ src/main/java/slaveFiles/Ui.java
@@ -1,0 +1,44 @@
+package slaveFiles;
+
+import java.util.List;
+
+public class Ui {
+    private String user = "slave driver";
+    private boolean hasMoreInputs;
+    private Parser parser;
+
+    public Ui(List<Task> tasklist, Storage activeStorage) {
+        parser = new Parser(tasklist, activeStorage);
+        hasMoreInputs = true;
+    }
+
+    public void getUserInputs(Storage storage) {
+        do {
+            parser.getUserInput(hasMoreInputs);
+        } while (hasMoreInputs);
+    }
+
+    /**
+     * Prints the greeting message
+     */
+    public void welcome() {
+        Parser.pageBreakLine();
+        // @@author Fluffykan-reused
+        // logo sourced from: https://www.askapache.com/online-tools/figlet-ascii/
+        String logo
+                = "_______        _______ _    _ _______\n"
+                + "|______ |      |_____|  \\  /  |______\n"
+                + "______| |_____ |     |   \\/   |______\n";
+        System.out.println("Ugh... why did you wake me up...\nGuess I am now your personal\n" + logo);
+        System.out.println("What do you want from me? Say it now, I don't have all day...");
+        Parser.pageBreakLine();
+    }
+
+    /**
+     *  Prints the goodbye message in response to the user inputting "bye"
+     */
+    public void goodbye() {
+        System.out.println("Good riddance " + user + ", try not to bother me in the future...");
+        Parser.pageBreakLine();
+    }
+}
